feat: 로컬 스토리지에 선택된 항목이 없다면 장바구니의 모든 아이템을 선택하는 기능 추가
